requirements.txt
asgiref==3.7.2
Django==4.2.7
django-crontab==0.7.1
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
phonenumbers==8.13.26
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlparse==0.4.4
